+++ .gitignore
@@ -1,0 +1,59 @@
+# Dependency directories
+/node_modules
+/.pnp
+.pnp.js
+
+# Output of build tools
+/build
+/dist
+
+# Environment variables
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# Log files
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+pnpm-debug.log*
+
+# Editor directories and files
+.vscode/
+.idea/
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+
+# macOS
+.DS_Store
+
+# Windows
+Thumbs.db
+
+# Visual Studio Code
+.vscode/
+
+# npm
+npm-debug.log*
+*.log
+
+# Yarn
+yarn-error.log*
+
+# ESLint
+.eslintcache
+
+# local .env files
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
+# Parcel
+.cache/
+dist/
+.cache
+.happypack/
