--- midfild/src/components/Results.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const Results = () => {
-    return (
-      <div id="Results" className='results-section'>
-          <div>
-            <h1>Results</h1>
-            <div>
-                <p> Results here </p>
-          </div>
-        </div>
-      </div>
-    );
-}
-
-export default Results;
